[power] Prepare node config targets for bootfs
Update each node config-data target to also build resource targets that
will include the node config file in bootfs. This is a no-op initially
but once the node config target is included as a dependency for a ZBI
target (coming in a later CL), the config file will be available to
power-manager in bootfs.

Test: first pave changes then `fx test power-manager-tests`
Test: Verify Astro boots and power-manager is running
Bug: fxbug.dev/65893
Change-Id: I0e25f469ea319833d8fa0b3a66a826eacb3f867a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460454
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>