[build] Fix ubsan variant build.
The "ubsan" is broken. The main issue is that libclang_rt.ubsan_standalone.so also
depends on libc++abi.so, and indirectly on libunwind.so, and this
is not properly reflected by the runtime.json file provided with or
Clang prebuilt toolchain.

This fixes the issue by patching runtime.json after reading it,
similarly to what is done for libclang_rt.asan.so which has a similar
issue.

Bug: 66509
Testing: "fx set core.x64 --goma --variant ubsan" then "fx build" should succeed.
Change-Id: I2885934c701318d92de5107bd44f412a0ca2c386
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463835
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: David Turner <digit@google.com>