[lowpan][lowpan-spinel-driver] Bring interface online if provisioned
Currently, if the LoWPAN device is provisioned and the system reboots,
the interface will remain in the disabled state, requiring someone to
do the equivalent of a `ifconfig lowpan0 up` to bring the interface
online.

This change fixes this situation by changing the default initial state
of the interface to enabled. In the course of implementing this fix,
several other issues were identified and corrected, such as some
confusion between the `enabled/disabled` and `online/offline` states
in `tun.rs` (and some other parts). The way addresses propagate from
the NCP to netstack was also made to be more robust and to also
exclude the internal mesh-local addresses that are not for application
use.

Test: fx test lowpan-tests
Change-Id: I1a44bc3a545a74a9b048faf6b83f64c0ebdfad47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460855
Fuchsia-Auto-Submit: Robert Quattlebaum <rquattle@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Testability-Review: Jiaming (Charlie) Wang <jiamingw@google.com>