[settings] Replace SettingResponse with SettingInfo
This CL completes the replacement of SettingResponse, which was part of
the switchboard, with SettingInfo, which resides in the base of the
service. The original CL, fxr/460716, duplicated the enum as part of a
larger effort to refactor the changed notifications so it didn't want
to make huge changes in the system. This CL finishes the move + rename.

Fixed: 66338
Test: fx test setui_service_tests
Change-Id: Iefdce4e04acce112d7140a022ddebebbe118ac33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463254
Commit-Queue: William Xiao <wxyz@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>