[weave] Remove ThreadStackManager Initialization
In some situations, the Thread radio is not responding, causing Weave to
fail initialization due to TSM failing its init. This causes WeaveStack
to exit in a failure condition, denying Weave and Weave-related
services. This initialization will be revisited, possibly with retries
or a softer failure mode.

Bug: 59029
Test: fx test weavestack-adaptation-unittests
Test: fx test weavestack-unittests
Change-Id: Icccdc98da44fd5c4ac6ea4453f499afb3d962fbc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463953
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Commit-Queue: Jason Graffius <jgraff@google.com>