[component_manager] Enforce Component V2 Capability Allowlist Policy.
This change enforces capability allowlist policy enforcement in Component
Framework V2. This change will prevent capabilities that are restricted
within the RuntimeConfiguration from being routed to realms that are not
defined in that set. Routing will only be blocked if a capability
attempting to be routed falls within that set.

Fixed: 61864
Change-Id: I5b8e44e5d6c3fab5238b9557d915689fe7930e42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460157
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Shai Barack <shayba@google.com>