[build][concurrency] properly track if using thinlto or lto
The use_thinlto variable is only active if use_lto is also set.

The logic in the concurrency computation has been wrong for a while
in this regard but it hasn't mattered until relatively recently
when local toolchain jobs have been tossed into the local pool when not
using goma.

Change-Id: I1e7612dbb24ed79854623c47b84fdffe37658c3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464035
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>