[zircon][kernel] iovec support for channels
This CL implements the kernel side of RFC-0010. Please refer to that RFC
for higher-level questions regarding motivations and alternatives.

When the ZX_CHANNEL_WRITE_USE_IOVEC option is provided to the
zx_channel_{write,write_etc,call,call_etc} syscalls, the `bytes`
argument is reinterpreted as an array of zx_channel_iovec_t, which refer
to multiple userspace byte arrays that should be copied into the message
buffer.

The BufferChain is updated in this CL to support allocating more pages
than are needed and freeing the unused ones before sending the message.
This has the advantage of avoiding multiple allocations of pages in the
case where the size of the final buffer is initially unknown.

Bug: 65677

Change-Id: I112339f449efbed629e09d6e3ce4ae9aae559f88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448877
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>