[vfs] Improve vfs header file.
Move OpenResult definition out of the VFS class to make the Vfs class
easier to follow. The OpenResult structure is the same.

Rename vdircookie_t to VdirCookie according to the style guide, and give
the members default values. Use the constructor rather than the custom
Reset method, which is now deleted.

Fix some undefined behavior in the microbenchmarks test. In two places
it forgot to initialize the vdir_cookie and depended on the value of the
dir cookie not changing between loop iterations even though the variable
was declared inside the loop. The variable is now properly initialized
the first time outside of the loop.

There should be no behavior change other than the test fix.

Change-Id: Ib372b6adec443bc4766fed825087bab38696b0c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463962
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Chris Suter <csuter@google.com>