[omaha-client] Add metrics tracking lost events
Update events are sent back to Omaha by the client, but this is on a
"best-effort" basis. Any issues resulting in the event being interrupted
or ignored do not result in retry. This change sends these updates to
our metrics reporter, which will ensure delivery of this information.

Added a new test to specifically track the production of this kind of
event, and updated an existing test that required it.

This must not be submitted until fxr/457834 is complete.

Test: fx test omaha-client-lib-tests
Test: fx test omaha-client-service-bin-tests
Fixed: 60594
Change-Id: I0534cc6554e6987f6695f18d632943584e6759d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458218
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Testability-Review: Aaron Wood <aaronwood@google.com>