[time] Remove fuchsia.time.Utc FIDL API.
This interface has been deprecated for a while and the final consumer
was recently removed, hence we can remove the API and implementation.

Note that the API was never included in the SDK hence this is not an SDK
change.

Bug: 60642
Change-Id: I978af5a2f81fd82c7660edec587f2480d02b7dbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464540
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>
API-Review: Jody Sankey <jsankey@google.com>