[magma][msd-vsi-vip] Increase hang check timeout.
To be able to complete the OpenVX-cts on some of the slower
devices the hang check timeout need to be increased.

The current timeout is 5000ms. On the slower devices we need a
6000ms for the tests to be able to run successfully.

Tested using OpenVX-cts on device.

[ ALL DONE ] 7342 test(s) from 81 test case(s) ran
[ PASSED   ] 7342 test(s)
[ FAILED   ] 0 test(s)
[ DISABLED ] 6985 test(s)

Full details of the testing are in the bug.

Bug: 66894
Change-Id: I7e9724605f931c4aa33835a409bb83208bf4a082
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464974
Commit-Queue: Jamie Howarth <jhowarth@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>