[check-licenses] Reorganizing textextensions list.
strictTextExtensionList in the config file is a list of filetypes that
should contain copyright information at the top of the file.

textExtensionList is a list of filetypes that can't explicitly contain
copyright info at the top, but they should still be covered by a valid
project-level license.

Logically, the strict list should be a subset of the other list, but
previously they were handled as entirely separate lists. This removes
the duplicate entries from the textExtensionList, and modifies the code
to do the right thing.

Change-Id: I78b3159083dfb9144e9b85932ee47623ed9353e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465574
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Sean Cuff <seancuff@google.com>