[ffx] update IPv6 scopes as they change
It is important to try to preserve scope information, which is tricky
because it is host-local. ffx will discover un-scoped addresses from
sources such as netstack queries over Overnet. In those cases, it is
best to preserve any previously discovered scope for the given
address, which likely came from a host-side discovery protocol such
as mdns. If the incoming merge has scope information, that
information should always be preferred, as it may be more fresh, such
as in the situation where a device restart triggers the interface id
to change.

Change-Id: I4340779283df1eb01ef0043357b6be8c9096d014
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465478
Commit-Queue: James Tucker <raggi@google.com>
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>