[fx host-tool] ensure build outputs to terminal
As host-tool is often used in a captured output capacity, the output
from a build would be disruptive. In order to avoid this problem, the
tool now checks to see if there is a controlling terminal, and if so,
redirect output to it. In the case where there is no controlling
terminal, or no terminal on stderr, the automatic build is skipped,
and an error is printed as if --no-build was passed.

Change-Id: I450b28778eca5927b4f1cbabb843fa31984aba1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465402
Commit-Queue: James Tucker <raggi@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>