[botanist] Clean up GCE target logs
The createInstance error logging on retry was making the logs hard to
read, especially because the invocation and its stderr are dumped to the
stdio of the task. Also log success to clearly indicate when the retries
completed.

Bug: 9127
Change-Id: Ifa6902f13085eea6941d9396bb2c8d5aabf69012
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465754
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>