[build] variant_toolchain_suite(): Add prefix_configs parameter
This allows one toolchain definition to provide a list of configs
that will be inserted before the list of default configs that apply
to all linkable targets in a given toolchain() instance.

This is required to move the kernel to the Fuchsia build, in order
to let it override system C and C++ headers with its own versions.

Bug: 66830
Testing: Does not change the build.
Change-Id: Ifa7acfa9d8c8355f67071d1d55ce06786b6d267b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465978
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>