[check-licenses] Reorganizing testdata directory.
filetree tests all need two files to be present:
 - want.json (representing our expected filetree)
 - config.json (used to generate the real filetree)

Previously I would have to dirty up the root "filetree" test directory
with these files for every test, but that would get difficult to
maintain. This CL moves those files into a subdirectory with the same
name as the test, and creates a new subdirectory called "root" which is
used as the basedir for the filetree being tested.

Change-Id: I0279f39e879d5728258025d5aea5879f05320045
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465575
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>