[kernel][x86] Additional assert in page table unmap
Adds an additional assert that ensures that the page we return to the
pmm is definitely the page at the entry that got unmapped. Although
this should be the case, there is non-trivial code between original
physical address calculation and entry modification that could contain
bugs.

Bug: 66978
Change-Id: Id5b7f06f214857a8e9b286f8b64891942b0ae64e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466056
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>