[system-update-checker] Prevent installing updates while current slot is pending
In order to guarantee the system is healthy during an OTA, we need to block OTAs on the
current system being committed. While in the update flow, we check if an
update's CommitStatus is Pending via polling CommitStatusProviderProxy.
Timeouts are interpreted as a pending commit.
If so, do not start a new update until the CommitStatus is set to Committed.

Added unit tests and integration tests which verify
behavior.

Fixed: 64603

Change-Id: I82bb9a61f697147fb984832eaecaa03d03368d44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460924
Commit-Queue: Olivia Nordquist <nolivia@google.com>
Reviewed-by: Zach Kirschenbaum <zkbaum@google.com>