[kernel][x86] Assert that a page table is not free before walking it
We are tracking a potential use-after-free in the x86 MMU code; while
walking page tables, we've found pages in the VM_PAGE_FREE / pmm free
state. It is possible commit 5654d72 fixed it, but to gain confidence
in the fix, add asserts that would catch the same class of bugs
in non-kasan builds.

The asserts check:
* Before reading a page table entry from a page table in the Query
  and Unmap paths, check that the page table page is not free.
* When Querying down a page table, check that next page tables are
  not free.

We use the 'is not free' state rather than 'is mmu' since there are
some early boot page tables that are not marked as MMU pages.

Bug: 66978 kasan panic while running core tests
Change-Id: Iec8b0b395c93b02b8b8f127f68e8c15482dd5ea1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466579
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>