[botanist] Teach GCE targets to use existing VMs
Eventually we will want to move instance creation out of the task in
order to save time. This prepares the target for that by having it
attach SSH keys to an existing instance when a name and zone are passed
in by the calling recipe. The recipe changes have not been made, so this
should be a no-op.

Bug: 9127
Change-Id: I2297fcb169b8763990907be68327ad13bbe8b8d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466580
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>