[system-updater] Add a timeout to cobalt flushes
We've seen some test flakes in which we appear to hang on Cobalt
flushes, though thankfully this won't be an issue in OMCL-based setups
(because they allow a reboot with PerformPendingReboot as soon as we
enter the WaitToReboot state, which happens _before_ we flush Cobalt
metrics).

Add a timeout to that task to ensure system-updater makes progress,
even if we drop metrics.

Bug: 64452
Test: refactored cobalt flush into a separate function, added unit
test.
MULTIPLY: system-updater-bin-test

Change-Id: I48fc79a8d10c41a25bc9518ebd76cc8be8953057
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466635
Commit-Queue: John Wittrock <wittrock@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>