[timekeeper] Remove "NotRequired" correction condition.
This condition was originally useful when we were managing a graceful
transition from never updating the clock to sometimes updating the
clock.

Having to handle this condition makes the error bound calculation
more complex (because we might need to update the error bound even
though we didn't need to update the time but then what if there was
still a slew in progress?). NotRequired only occurs very rarely today
and there really isn't a downside to implementing a short slew in these
cases.

Bug: 63368
Change-Id: I033fe3df50a822d9b2c34ff2f2f0f413680e3047
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466173
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>