[component_manager] Reland Enforce Component V2 Capability Allowlist Policy.
This change enforces capability allowlist policy enforcement in Component
Framework V2. This change will prevent capabilities that are restricted
within the RuntimeConfiguration from being routed to realms that are not
defined in that set. Routing will only be blocked if a capability
attempting to be routed falls within that set.

This change fixes the flake in the storage capability testing.

Fixed: 61864
Change-Id: I2389154edf15ffbcc3fd3ecaaadb98e5a42f7be4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463191
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Gary Bressler <geb@google.com>