[check-licenses] Make FileTree.Children an ordered list.
The iteration order for maps is undefined. Let's keep this data in a list so we can ensure the results won't change between runs.

Also remove "OutputTreeStateFilename" functionality. I previously used it to quickly save the state of a filetree while debugging, but we should use template.go to do that moving forward.

Change-Id: Ib566d6d13c806982a7abd6ac356553178e65a726
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467365
Reviewed-by: Sean Cuff <seancuff@google.com>
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>