[tee_manager] Switch tests to use App protocol
This CL switches the optee_test and optee_smoke_test to use the
fuchsia.tee.Application protocol instead of fuchsia.tee.Device. It also
eliminates the enclosing environment for the tests, as that was already
accomplished by the injected services of the hermetic test components.

Bug: 44664
Change-Id: I0b4a08f0e44393bb9c809cd743200f715de88574
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463238
Commit-Queue: Jonathan Bailey <jonathanbailey@google.com>
Reviewed-by: RJ Ascani <rjascani@google.com>