[debugger] Clear debug_agent's states on disconnecting
This creates a more consistent behavior between the debug_agent and the
zxdb so that further reconnecting will work as expected.

This patch also allows a debug_agent to run as a deamon in the
background as multiple zxdb sessions could connect in sequence.

Test: disconnected & reconnected the same zxdb session multiple times.
      ran multiple zxdb sessions against the same debug_agent.

Fixed: 64872
Fixed: 60034
Change-Id: If01743e8641a5782c433c5c3d71ab0b2fdf748c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467734
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>