[carnelian] Experiment with easier rendering
This change introduces the experimental concepts of Scenes and Facets,
in order to remove a lot of complex boilerplate in rendering and to
provide a way to provide composition of user-interface components in
Carnelian apps.

In this CL, a scene contains a hash map of boxed object implementing
the Facet trait. There is a separate vector of facet IDs that maintains
the drawing order of facets. The scene also maintains a per-facet
mutable array of layers used to render that facet.

When a frame is being produced, each facet is given the opportunity to
modify its vector of layers. This design is expected to eventually keep
track of the modifications to each facet's layers so as to make the
minimal changes to the composition, but this CL instead replaces all
layers every frame.

In this iteration, there are no common attributes for every Facet. All
mutation of Facets goes through a Any-based message handling method.
Future iterations will pull out common attributes like colors and
locations so that they can be managed by the scene.

As an exploration of the possibilities of this approach, scenes can
optionally draw the mouse cursor when needed as well as rounding the
corners of the view.

The ViewAssistant trait has a new optional method, resize(), which is
used for Scene-based apps to discard their scenes to be rebuilt on the
next call to render. In future iterations we can experiment with more
interesting approachs to handling view size changes.

Included in this change is the first use of the ttf-parser crate, which
provides access to more font metrics than our existing font crates and
which allows for the kind of vertical centering desired for
user-interfaces.

This CL introduces a new sample, font_metrics, which was used both to
try out the ttf-parser features as well as a proving ground for Scene
and Facet.

This CL also ported button, clockface, shapes, spinning_square and svg
to use Scenes and Facets. spinning_square was enhanced to provide a way
to demonstrate and modify drawing order of facets; press 'f' and 'b' on
a keyboard while it is running to see.

The remaining samples use features of render that are unlikely to ever
be supported by Scene and Facet and are being left alone for now.

Change-Id: I2b938fffcabee26f0926342ad80b0b3f6d961a30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427460
Commit-Queue: Rob Tsuk <robtsuk@google.com>
Reviewed-by: Dragos Tiselice <dtiselice@google.com>
Reviewed-by: David Reveman <reveman@google.com>