[events] Add support for individual sync/async events.
This CL introduces the event mode property in manifest files which
currently consists of async and sync. Those properties can be routed
across events. They're a set (of currently a max of two items). As you
route an event, we take the intersection of that mode set each step of
the way. This signifies whether or not a given component can support
registering to async and/or sync events for a particular event name.

Additionally, this CL introduces the notion of an
EventSubscriptionRequest which lets individual clients specify whether
they want sync or async events for a given set of event names. A single
event stream can receive a combination of sync and async events but only
one of each for a given event name.

Change-Id: I19cfc98f87f128225a6b90e3df3043eaa5d79917
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/456617
Commit-Queue: Fady Samuel <fsamuel@google.com>
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>