[nelson][spi] Fix chip select GPIO name
The Amlogic SPI driver expects each GPIO fragment name to be in the form
"gpio-cs-n", where n is the chip index. The Nelson board driver was not
updated to use this naming scheme, so this change fixes it and adds a
device-enumeration test entry.

Test: runtests -n device-enumeration-test
Change-Id: Ideaf3208db833753396015d30f602bfa3232c2bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467474
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>