[timekeeper] Remove Clock::Now
zx_clock_get is being removed as part of RFC-0008. This removes the
Clock::Now method which uses zx_clock_get and the associated clock ids.
This is the last step in a series of changes to remove uses of
zx_clock_get from timekeeper.

Bug: 60521

Change-Id: Icc1308e0cdb329967d37780ebdb6840fd5c16def
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463259
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>