[ffx] Timeout identify_host call in OvernetPeer handler.
In cases where a target reboots after the daemon has opened an RCS
connection to it, the call to identify_host in
Target::from_rcs_connection can sometimes hang indefinitely. This
appears to be a cause of issue 66905.

Adding a timeout to this call prevents the hang. I'm so far unable to
repro issue 66905 with this change.

Bug: 66905
Change-Id: I8ba878694b63e94c50d797f72320dd6d79c227a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467203
Commit-Queue: Jordon Wing <jwing@google.com>
Fuchsia-Auto-Submit: Jordon Wing <jwing@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>