[ffx] Clean up 'ffx component' and add additional help text
The goal is to have a comprehensive help and examples so that the user
is able to understand the commands and operations.

Also provides some guidance for devs who are contributing plugins to
follow a similar structure and language for the help text.

Change-Id: Iec7338aa8f88fc0f51843254f6c386be7eaeca50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466856
Commit-Queue: Amit Uttamchandani <amituttam@google.com>
Fuchsia-Auto-Submit: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>