[launchpad] Check for overflow when using the value of p_filesz
elf_load_find_interp directly reports the length of the interpreter
name as the PT_INTERP p_filesz header value. Since this is technically
untrusted, it could be SIZE_MAX, and the addition of 1 (for the
trailing NUL) could overflow.

So, check for this condition.

Bug: fxbug.dev/47998
Change-Id: I1b8a021a8259377edffd832a472f306b6f3265c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458275
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: George Kulakowski <kulakowski@google.com>