[time][e2e] Use setupAll for SL4F setup
On occasion this test times out, and the retry fails
due to a "client closed" error. The SL4F client is
closed when teardown is called. This updates the test
to setup and teardown SL4F once for the whole set of
tests in an attempt to reduce instances of the error.
Since SL4F listens on a hardcoded port, the test cases
already share an instance of SL4F if run concurrently.

Bug: 67173

Multiply: host_x64/time_sync_test

Change-Id: I26fd3cdb90ae5921272f894e244d1013b072446f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467854
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>