[testsharder] Prefer less runs for tests listed in multiple modifiers.
A test may appear in an affected tests file which will cause it to be
run up to a default max number of times or as many times as will fit in
the target duration according to test duration data. However, this may
cause some issues in certain cases like go/bbid/8860185618054452560
where the affected test had many test cases and produced a really large
amount of logs that the builder failed due to running out of memory. In
that case, we should be able to explicitly set a smaller number of runs
through the MULTIPLY feature. However, that causes two multipliers to be
created for the same test (one from MULTIPLY and one from the affected
tests file). The current testsharder takes the number of runs associated
with the last multiplier in the list (the one from the affected tests).
This change will make it take the min number of runs from all
multiplers for the same test.

MULTIPLY: testsharder: 500
Change-Id: Ie8f7763973ecacf90e871829200dd5839e245ef6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465054
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>