[fx ota] wait for commit post-reboot
This change modifies `fx ota` so that post-reboot, it waits for the
update to commit. If the commit hangs for more than 30 seconds, a
message will print to recommend reverting the update with `update
revert`.

Test: fx self-test subcommands/fx_ota_test
Test: Manually observed it print. Also, added a timer to stall the
commit and observed the "hang for 30 seconds" printline.
Fixed: 64604

Change-Id: I69078c2ed7985d7cf798c206cde82638ee3e4448
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467197
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>