[timekeeper] Set error bounds duing clock updates.
This covers setting an error bound in all existing updates and making
additional clock updates mid-slew to reduce the error bound if
necessary. There will be a subsequent CL to increase the error bound if
no time samples are received for an extended period.

Bug: 63368
Change-Id: Ic008625d479535899d283dd460a2f48657a2bf08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466214
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>