[kernel][kasan] Quarantine pmm frees to maximize reuse distance
Kernel Address Sanitizer is a kernel memory error detector that
detects certain use-out-of-bounds and use-after-free errors via
compiler instrumentation and sideband information.

In order to detect use-after-free errors, address sanitizer would
prefer that the time between a buffer is freed and the time it
is reused is maximized; in the kernel heap, we "quarantine" freed
buffers for a short time before releasing them to the heap free
lists.

This CR modifies the PMM to insert freed pages at the tail, rather
than the head of the free list (the PMM allocates from the head of
the free list) in address sanitizer builds. Inserting at the tail
maximizes the time between free()ing a page and subsequent
reallocations of that page without additional memory use.

In the future, this feature may be desirable at runtime while the
PMM free-fill checker is active; defer that for now, as the
performance characteristics are not yet understood.

Bug: 30033
Pair-Programmed-With: mvanotti@google.com
Change-Id: Id2d6b0383d385c9c0451b5d7ed0d26274f4483e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467598
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>