[debugger] Handle deleting a target properly.
Previously when deleting the current target, it would pick the first
one. But this was done before the old one was deleted. The result was
that ConsoleContext would pick the same target that's about to be
deleted if the first one is deleted.

This reorders the logic to avoid this case.

Change-Id: Ic9b1e8943446efc9fa78504ec17518f67ec4c88c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468236
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>