[flake][time] Increase time e2e test timeout
While time synchronization itself is expected to complete within a
minute in most cases, on emulators the timeout can still be exceeded if
the device is busy. This seems to occur because the requires sl4f to be
started via ssh and to respond to http requests, which do not complete
in time.

Bug: 67173

Change-Id: Ibbd11d99ca4a21bc88a5073c67c27acd47db12be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468240
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>