[fx] Better message when goma is not started
If goma is enabled but not started, 'fx build' will present a generic
error message that instructs the user to run 'goma_ctl.py' instead of
the 'fx goma' wrapper. There is no way to make that message better
because it is generic for all goma customers.

This CL checks if goma is enabled in args.gn and then prints an
fx-specific error message if Goma is not started. This happens only
after a failed Ninja call, so that successful builds don't pay the price
to check goma status.

Fixed: 67059
Change-Id: I1ad8354e665e3a153d3d3768cfdebb1045608fd7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466694
Fuchsia-Auto-Submit: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Haowei Wu <haowei@google.com>